requirementss.txt
streamlit==1.33.0
tensorflow==2.12.0
Pillow==9.5.0
numpy==1.23.5
protobuf==3.20.3